change balance to Decimal

+++ accounts/models.py
@@ -6,7 +6,7 @@ class Account(models.Model):
      first_name = models.CharField(max_length=50)
      last_name = models.CharField(max_length=50)
      inn = models.BigIntegerField(db_index=True)
-     balance = models.FloatField(default=0.0)
+     balance = models.DecimalField(max_digits=10, decimal_places=2, default=0.0)
 
      def __str__(self):
          return self.first_name + ' ' + self.last_name + ' (' + str(self.inn) + ')' + ' (balance = ' + str(self.balance) + ')'
